edit some frames
