+++ src/MainApp.java
@@ -1,72 +1,7 @@
 package src;
 
-import java.awt.BorderLayout;
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.FlowLayout;
-import java.awt.GridLayout;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-import javax.swing.*;
-
-import src.AddressSearch.AddressSearchOption;
-import src.Knapsack.KnapsackMain;
-
-import src.SelectionSort.SelectionSort;
-import src.TSP.Tsp;
-
-public class MainApp extends JFrame implements ActionListener {
-
-    public MainApp() {
-        //Di malinis code ko dine yawa basta nasa iisang panel lang sya
-        JPanel startPanel = new JPanel();
-        startPanel.setLayout(new BorderLayout());
-        startPanel.setBounds(0, 0, 500, 500);
-
-        //Image 
-        ImageIcon attitudeIcon = new ImageIcon("resources/attitude.png");
-        JLabel attitudeLabel = new JLabel();
-        attitudeLabel.setIcon(attitudeIcon);
-        attitudeLabel.setPreferredSize(new Dimension(20, 20));
-        attitudeLabel.setBounds(0,500 ,500,500);
-        attitudeLabel.setVerticalAlignment(JLabel.CENTER);
-        attitudeLabel.setHorizontalAlignment(JLabel.CENTER);
-        
-        //Only one button only one yieee
-        JButton start = new JButton("Start");
-        start.setBounds(425, 370, 100, 100);
-        start.addActionListener(this);
-        
-        //add to the panel
-        startPanel.add(start,BorderLayout.CENTER);
-        startPanel.add(attitudeLabel,BorderLayout.CENTER);
-        setTitle("Midterm Project");
-        setDefaultCloseOperation(EXIT_ON_CLOSE);
-        setSize(1000, 563);
-        setVisible(true);
-        setResizable(false);
-        //add into the frame
-        this.add(startPanel);
-        
-        
-        
-
-    }
-
+public class MainApp {
     public static void main(String[] args) {
-        new MainApp();
-
-         //KnapsackMain knapsackMain = new KnapsackMain();
-         //SelectionSort selectionSort = new SelectionSort();
-        //Tsp tsp = new Tsp();
-         //AddressSearchOption addressSearchOption = new AddressSearchOption();
-
+        Frame1 start = new Frame1();
     }
-
-    public void actionPerformed(ActionEvent e) {
-        dispose();
-        Frame2 frame2 = new Frame2();
-    }
-
 }
